Move Plug entry point to ImagePipe.Plug
